Added river and adjusted to implement color change
I've added the river shape definition, drawRiver, and fillShapeWithHorizontalLines functions back into the sketch. I did this first as the river shares the same color array as the sky, so it was easier to implement the new changing color array into this section of the artwork. In the drawRiver function, I replaced the skyColor array with the changingSkyColor array. I also had to initialise the changingSkyColor array in the initialiseData section to ensure that it was accessible globally.