Added buildings/reflection and adjusted to implement color change
I've used the same logic to initialise a second building color array and an empty changing building color array to be implemented in the building and its reflection. I've added the building/reflection shape definitions, drawBuilding, drawBuilding2, drawReflection, and fillShapeWithVerticalLines back into the code. I've adjusted the drawBuilding and drawReflection code in a similar manner to the river, but this time implementing the changing building color array that interacts with the mouse x-position.